+++ CHANGELOG.md
@@ -13,6 +13,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [2.1.6] - 2026-03-09
+### Added
+- AllaganLib.GameSheets
+  - LuminaSupplemental.Excel updated to 4.2.3
+  - All Relic Weapons and Relic Tools are now provided as sources
+
+### Changed
+- AllaganLib.GameSheets
+  - ILocation now provided a HasCoordinates indicating if it's a general map location or a specific location 
+
+## [2.1.5] - 2026-03-09
+### Added
+- AllaganLib.GameSheets
+  - LuminaSupplemental.Excel updated to 4.2.2
+  - QuestExtensions added
+  - BGMOrchestrion added to SheetManager
+
 ## [2.1.4] - 2026-03-09
 ### Added
 - AllaganLib.GameSheets